fix: tsconfig.json presets
