+++ native/src/screens/pages/SignIn/styles.ts
@@ -1,0 +1,58 @@
+const estilos = StyleSheet.create({
+    container: {
+      padding: 30,
+      width: '100%',
+      paddingBottom: 48,
+      backgroundColor: '#fff',
+    },
+    label: {
+      fontSize: 16,
+      marginTop: 16,
+      marginBottom: 4,
+    },
+    input: {
+      borderWidth: 1,
+      borderColor: '#999',
+      borderRadius: 4,
+      padding: 12,
+    },
+    picker: {
+      borderWidth: 1,
+      borderColor: '#999',
+      borderRadius: 4,
+      overflow: 'hidden',
+    },
+    classeContainer: {
+      flexDirection: 'row',
+      justifyContent: 'space-around',
+      marginVertical: 12,
+    },
+    botaoClasse: {
+      paddingVertical: 10,
+      paddingHorizontal: 20,
+      backgroundColor: '#ddd',
+      borderRadius: 6,
+    },
+    classeAtiva: {
+      backgroundColor: '#000',
+    },
+    textoBotaoClasse: {
+      color: '#fff',
+      fontWeight: 'bold',
+    },
+    botaoFinal: {
+      backgroundColor: '#4B4DED',
+      padding: 16,
+      borderRadius: 6,
+      marginTop: 24,
+    },
+    textoBotaoFinal: {
+      color: '#fff',
+      textAlign: 'center',
+      fontWeight: 'bold',
+      fontSize: 16,
+    },
+  });
+
+
+  export default estilos
